+++ packages/builder/builder-rspack-provider/src/core/rspackConfig.ts
@@ -93,6 +93,7 @@ async function getChainUtils(target: BuilderTarget): Promise<ChainUtils> {
     target,
     isProd: nodeEnv === 'production',
     isServer: target === 'node',
+    isServiceWorker: target === 'service-worker',
     isWebWorker: target === 'web-worker',
     getCompiledPath,
     CHAIN_ID,