+++ packages/builder/builder-rspack-provider/src/plugins/hmr.ts
@@ -8,7 +8,12 @@ import { setConfig } from '@modern-js/builder-shared';
 export const isUsingHMR = (
   config: NormalizedConfig,
   { isProd, target }: ModifyRspackConfigUtils,
-) => !isProd && target !== 'node' && target !== 'web-worker' && config.dev.hmr;
+) =>
+  !isProd &&
+  target !== 'node' &&
+  target !== 'web-worker' &&
+  target !== 'service-worker' &&
+  config.dev.hmr;
 
 export const builderPluginHMR = (): BuilderPlugin => ({
   name: 'builder-plugin-hmr',