+++ packages/builder/builder-rspack-provider/src/shared/plugin.ts
@@ -13,9 +13,10 @@ export const applyMinimalPlugins = (plugins: Plugins) =>
     plugins.fileSize(),
     // should before the html plugin
     plugins.cleanOutput(),
-    import('../plugins/font').then(m => m.builderPluginFont()),
-    import('../plugins/image').then(m => m.builderPluginImage()),
-    import('../plugins/media').then(m => m.builderPluginMedia()),
+    plugins.font(),
+    plugins.image(),
+    plugins.media(),
+    plugins.svg(),
     import('../plugins/html').then(m => m.builderPluginHtml()),
     import('../plugins/define').then(m => m.builderPluginDefine()),
     import('../plugins/css').then(m => m.builderPluginCss()),
@@ -36,4 +37,5 @@ export const applyDefaultPlugins = (plugins: Plugins) =>
     plugins.yaml(),
     plugins.splitChunks(),
     plugins.startUrl(),
+    plugins.bundleAnalyzer(),
   ]);