+++ packages/cli/doc-core/src/node/searchIndex.ts
@@ -1,7 +1,9 @@
 import path, { join, resolve } from 'path';
 import fs from '@modern-js/utils/fs-extra';
+import chalk from '@modern-js/utils/chalk';
 import { RequestHandler } from '@modern-js/types';
-import { OUTPUT_DIR, PUBLIC_DIR } from './constants';
+import fetch from 'node-fetch';
+import { isProduction, OUTPUT_DIR, PUBLIC_DIR } from './constants';
 import { UserConfig } from '@/shared/types';
 import { addLeadingSlash, SEARCH_INDEX_JSON } from '@/shared/utils';
 
@@ -12,6 +14,30 @@ export async function writeSearchIndex(rootDir: string, config: UserConfig) {
   const target = join(cwd, OUTPUT_DIR, 'static', SEARCH_INDEX_JSON);
   if (await fs.pathExists(source)) {
     await fs.move(source, target, { overwrite: true });
+    if (isProduction() && config.doc?.search?.mode === 'remote') {
+      const { apiUrl, indexName } = config.doc.search;
+      const indexData = await fs.readFile(target);
+      try {
+        await fetch(`${apiUrl}?index=${indexName}`, {
+          method: 'PUT',
+          body: indexData,
+          headers: { 'Content-Type': 'application/json' },
+        });
+        // eslint-disable-next-line no-console
+        console.log(
+          chalk.green(
+            `[doc-tools] Search index uploaded to ${apiUrl}, indexName: ${indexName}`,
+          ),
+        );
+      } catch (e) {
+        // eslint-disable-next-line no-console
+        console.log(
+          chalk.red(
+            `[doc-tools] Upload search index \`${indexName}\` failed:\n ${e}`,
+          ),
+        );
+      }
+    }
   }
 }