+++ packages/cli/doc-core/src/node/virtualModule/siteData.ts
@@ -6,6 +6,7 @@ import {
   SidebarItem,
   SidebarGroup,
   NormalizedSidebarGroup,
+  PageIndexInfo,
 } from 'shared/types';
 import VirtualModulesPlugin from 'webpack-virtual-modules';
 import { remark } from 'remark';
@@ -16,7 +17,7 @@ import { htmlToText } from 'html-to-text';
 import remarkParse from 'remark-parse';
 import remarkHtml from 'remark-html';
 import { remarkPluginContainer } from '@modern-js/remark-container';
-import { ReplaceRule, Header } from 'shared/types/index';
+import { ReplaceRule } from 'shared/types/index';
 import fs from '@modern-js/utils/fs-extra';
 import { parseToc } from '../mdx/remarkPlugins/toc';
 import { importStatementRegex, PACKAGE_ROOT, PUBLIC_DIR } from '../constants';
@@ -25,20 +26,11 @@ import { flattenMdxContent } from '../utils/flattenMdxContent';
 import { routeService } from './routeData';
 import { MDX_REGEXP, withBase } from '@/shared/utils';
 
-interface PageIndexData {
-  id: number;
-  title: string;
-  routePath: string;
-  toc: Header[];
-  content: string;
-  frontmatter: Record<string, unknown>;
-}
-
-let pages: PageIndexData[] | undefined;
+let pages: PageIndexInfo[] | undefined;
 
 export function normalizeThemeConfig(
   themeConfig: DefaultThemeConfig,
-  pages: PageIndexData[] = [],
+  pages: PageIndexInfo[] = [],
   base = '',
   replaceRules: ReplaceRule[],
 ): NormalizedDefaultThemeConfig {
@@ -113,7 +105,8 @@ export function normalizeThemeConfig(
 async function extractPageData(
   replaceRules: ReplaceRule[],
   alias: Record<string, string | string[]>,
-): Promise<PageIndexData[]> {
+  domain: string,
+): Promise<(PageIndexInfo | null)[]> {
   return Promise.all(
     routeService
       .getRoutes()
@@ -124,6 +117,9 @@ async function extractPageData(
           // eslint-disable-next-line import/no-named-as-default-member
           ...yamlFront.loadFront(content),
         };
+        if (frontmatter.pageType === 'home') {
+          return null;
+        }
         // 1. Replace rules for frontmatter & content
         Object.keys(frontmatter).forEach(key => {
           if (typeof frontmatter[key] === 'string') {
@@ -146,6 +142,9 @@ async function extractPageData(
         // 2. Optimize content index
         const ast = remark.parse({ value: content });
         const { title, toc } = parseToc(ast as Root);
+        if (!title?.length && !frontmatter.title?.length) {
+          return null;
+        }
         const precessor = unified()
           .use(remarkParse)
           .use(remarkPluginContainer)
@@ -164,18 +163,29 @@ async function extractPageData(
               selector: 'img',
               format: 'skip',
             },
+            ...['h1', 'h2', 'h3', 'h4', 'h5', 'h6'].map(tag => ({
+              selector: tag,
+              options: {
+                uppercase: false,
+              },
+            })),
           ],
-          uppercaseHeadings: false,
           tables: true,
           longWordSplit: {
             forceWrapOnLimit: true,
           },
         });
+        if (content.startsWith(title)) {
+          // Remove the title from the content
+          content = content.slice(title.length);
+        }
         return {
           id: index,
           title: frontmatter.title || title,
           routePath: route.routePath,
+          lang: route.lang,
           toc,
+          domain,
           // Stripped frontmatter content
           content,
           frontmatter: {
@@ -201,8 +211,14 @@ export async function siteDataVMPlugin(
     console.log('⭐️ [doc-tools] Extracting site data...');
   }
   const replaceRules = userConfig?.replaceRules || [];
+  const domain =
+    userConfig?.search?.mode === 'remote'
+      ? userConfig?.search.domain ?? ''
+      : '';
   if (!pages) {
-    pages = await extractPageData(replaceRules, alias);
+    pages = (await extractPageData(replaceRules, alias, domain)).filter(
+      Boolean,
+    ) as PageIndexInfo[];
   }
   const siteData = {
     title: userConfig?.title || '',
@@ -218,6 +234,7 @@ export async function siteDataVMPlugin(
     root: userRoot,
     lang: userConfig?.lang || 'zh',
     logo: userConfig?.logo || '',
+    search: userConfig?.search || { mode: 'local' },
     pages: pages.map(({ routePath, toc }) => ({
       routePath,
       toc,