+++ scripts/prebundle/src/constant.ts
@@ -211,12 +211,52 @@ export const TASKS: TaskConfig[] = [
           'loader-utils': '../loader-utils2',
         },
       },
+      {
+        name: 'schema-utils3',
+        ignoreDts: true,
+      },
+      {
+        name: 'url-loader',
+        ignoreDts: true,
+        externals: {
+          'schema-utils': '../schema-utils3',
+          'loader-utils': '../loader-utils2',
+          'mime-types': '@modern-js/utils/mime-types',
+        },
+        afterBundle(task) {
+          replaceFileContent(join(task.distPath, 'index.js'), content => {
+            // using prebunled file-loader
+            return content.replace(
+              '"file-loader"',
+              'require.resolve("../file-loader")',
+            );
+          });
+        },
+      },
+      {
+        name: 'file-loader',
+        ignoreDts: true,
+        externals: {
+          'schema-utils': '../schema-utils3',
+          'loader-utils': '../loader-utils2',
+        },
+      },
     ],
   },
   {
     packageDir: 'builder/builder-shared',
     packageName: '@modern-js/builder-shared',
-    dependencies: ['open', 'webpack-5-chain'],
+    dependencies: [
+      'open',
+      'webpack-5-chain',
+      {
+        name: 'webpack-bundle-analyzer',
+        externals: {
+          chalk: '@modern-js/utils/chalk',
+          'gzip-size': '@modern-js/utils/gzip-size',
+        },
+      },
+    ],
   },
   {
     packageDir: 'builder/builder-webpack-provider',
@@ -239,13 +279,6 @@ export const TASKS: TaskConfig[] = [
       {
         name: 'webpack-sources',
         ignoreDts: true,
-      },
-      {
-        name: 'loader-utils2',
-        ignoreDts: true,
-        externals: {
-          json5: '@modern-js/utils/json5',
-        },
       },
       {
         name: 'schema-utils3',
@@ -389,39 +422,6 @@ export const TASKS: TaskConfig[] = [
         name: 'sass-loader',
         externals: {
           sass: '../sass',
-        },
-      },
-      {
-        name: 'url-loader',
-        ignoreDts: true,
-        externals: {
-          'schema-utils': '../schema-utils3',
-          'loader-utils': '../loader-utils2',
-          'mime-types': '@modern-js/utils/mime-types',
-        },
-        afterBundle(task) {
-          replaceFileContent(join(task.distPath, 'index.js'), content => {
-            // using prebunled file-loader
-            return content.replace(
-              '"file-loader"',
-              'require.resolve("../file-loader")',
-            );
-          });
-        },
-      },
-      {
-        name: 'file-loader',
-        ignoreDts: true,
-        externals: {
-          'schema-utils': '../schema-utils3',
-          'loader-utils': '../loader-utils2',
-        },
-      },
-      {
-        name: 'webpack-bundle-analyzer',
-        externals: {
-          chalk: '@modern-js/utils/chalk',
-          'gzip-size': '@modern-js/utils/gzip-size',
         },
       },
       {