+++ packages/builder/builder-rspack-provider/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @modern-js/builder-rspack-provider
 
+## 2.4.0
+
+### Minor Changes
+
+- a5572b8: feat: more plugin rebase in `builder-shared`
+  feat: 更多 plugin 下沉到 `builder-shared`
+- a914be8: feat: modern-js support rspack bundler
+  feat: modern-js 支持 rspack 构建工具
+
+### Patch Changes
+
+- 6f83037: fix: change the builder resolve plugin
+  fix: 修复 builder resolve 插件
+- 014d06b: feat: reuse bundleAnalyzer plugin, support performance.bundleAnalyze config in rspack-provider
+
+  feat: 复用 bundleAnalyzer 插件，在 rspack-provider 中支持 performance.bundleAnalyze 配置项
+
+- 48b036e: fix(builder): should not generate cache when build failed
+
+  fix(builder): 修复构建失败时会生成无效编译缓存的问题
+
+- 8c2db5f: feat(core): improve support for exporting a function in config file
+
+  feat(core): 完善对配置文件中导出函数的支持
+
+- Updated dependencies [d19484c]
+- Updated dependencies [6f83037]
+- Updated dependencies [014d06b]
+- Updated dependencies [98a2733]
+- Updated dependencies [a5572b8]
+- Updated dependencies [a914be8]
+- Updated dependencies [8c2db5f]
+  - @modern-js/server@2.4.0
+  - @modern-js/builder-shared@2.4.0
+  - @modern-js/utils@2.4.0
+  - @modern-js/e2e@2.4.0
+  - @modern-js/types@2.4.0
+
 ## 2.3.0
 
 ### Patch Changes